.ipynb_checkpoints/requirements-checkpoint.txt
pandas==1.1.3
streamlit==0.77.0
Pillow==8.2.0
